+++ assets/js/web-components/prpl-gauge.js
@@ -1,10 +1,10 @@
-/* global customElements, HTMLElement */
+/* global customElements, HTMLElement, prplUpdatePreviousMonthBadgeProgressBar */
 /*
  * Web Component: prpl-gauge
  *
  * A web component that displays a gauge.
  *
- * Dependencies: progress-planner/web-components/prpl-badge
+ * Dependencies: progress-planner/web-components/prpl-badge, progress-planner/web-components/prpl-badge-progress-bar
  */
 
 /**
@@ -174,16 +174,16 @@ const prplUpdateRaviGauge = ( pointsDiff ) => {
 	// Mark badge as completed, in the a Monthly badges widgets, if we reached the max points.
 	if ( newValue >= parseInt( gaugeProps.max ) ) {
 		// We have multiple badges, one in widget and the other in the popover.
-		const badges = document.querySelectorAll(
-			'.prpl-badge-row-wrapper-inner .prpl-badge prpl-badge[complete="false"][badge-id="' +
-				gaugeProps.badgeId +
-				'"]'
-		);
-
-		if ( badges ) {
-			badges.forEach( ( badge ) => {
+		document
+			.querySelectorAll(
+				`.prpl-badge-row-wrapper-inner .prpl-badge prpl-badge[complete="false"][badge-id="${ gaugeProps.badgeId }"]`
+			)
+			?.forEach( ( badge ) => {
 				badge.setAttribute( 'complete', 'true' );
 			} );
+
+		if ( gaugeProps.value >= parseInt( gaugeProps.max ) ) {
+			prplUpdatePreviousMonthBadgeProgressBar( pointsDiff );
 		}
 	}
 };