+++ classes/badges/class-monthly.php
@@ -122,6 +122,30 @@ final class Monthly extends Badge {
 	}
 
 	/**
+	 * Get the badge object from a badge ID.
+	 *
+	 * @param string $badge_id The badge ID.
+	 *
+	 * @return \Progress_Planner\Badges\Monthly|null
+	 */
+	public static function get_instance_from_id( $badge_id ) {
+		$year  = (int) explode( '-', str_replace( 'monthly-', '', $badge_id ) )[0];
+		$month = (int) str_replace( 'm', '', explode( '-', str_replace( 'monthly-', '', $badge_id ) )[1] );
+
+		$instances = self::get_instances();
+
+		if ( isset( $instances[ $year ] ) ) {
+			foreach ( $instances[ $year ] as $instance ) {
+				if ( (int) $instance->get_month() === $month ) {
+					return $instance;
+				}
+			}
+		}
+
+		return null;
+	}
+
+	/**
 	 * Get an array of months.
 	 *
 	 * @return array
@@ -189,13 +213,19 @@ final class Monthly extends Badge {
 	/**
 	 * Progress callback.
 	 *
+	 * @param array $args The arguments for the progress callback.
+	 *
 	 * @return array
 	 */
-	public function progress_callback() {
+	public function progress_callback( $args = [] ) {
 		$saved_progress = $this->get_saved();
 
 		// If we have a saved value, return it.
-		if ( isset( $saved_progress['progress'] ) && isset( $saved_progress['remaining'] ) ) {
+		if ( isset( $saved_progress['progress'] )
+			&& isset( $saved_progress['remaining'] )
+			&& isset( $saved_progress['points'] )
+			&& 100 === $saved_progress['progress']
+		) {
 			return $saved_progress;
 		}
 
@@ -220,17 +250,65 @@ final class Monthly extends Badge {
 			$points += $activity->get_points( $activity->date );
 		}
 
-		$return_progress = ( $points > self::TARGET_POINTS )
-			? [
-				'progress'  => 100,
-				'remaining' => 0,
-			] : [
-				'progress'  => (int) max( 0, min( 100, floor( 100 * $points / self::TARGET_POINTS ) ) ),
-				'remaining' => self::TARGET_POINTS - $points,
-			];
+		$return_progress = [
+			'progress'  => (int) max( 0, min( 100, floor( 100 * $points / self::TARGET_POINTS ) ) ),
+			'remaining' => (int) max( 0, min( self::TARGET_POINTS - $points, 10 ) ),
+			'points'    => $points,
+		];
 
 		$this->save_progress( $return_progress );
 
-		return $return_progress;
+		if ( $points >= self::TARGET_POINTS || ( isset( $args['no_next_badge_points'] ) && $args['no_next_badge_points'] ) ) {
+			return $return_progress;
+		}
+
+		$points += $this->get_next_badges_excess_points();
+		return [
+			'progress'  => (int) max( 0, min( 100, floor( 100 * $points / self::TARGET_POINTS ) ) ),
+			'remaining' => (int) max( 0, min( self::TARGET_POINTS - $points, 10 ) ),
+			'points'    => $points,
+		];
+	}
+
+	/**
+	 * Get the next badge-ID.
+	 *
+	 * @return string
+	 */
+	public function get_next_badge_id() {
+		$year     = $this->get_year();
+		$month    = $this->get_month();
+		$month    = $month < 10 ? "0$month" : $month;
+		$datetime = \DateTime::createFromFormat( 'Y-m-d', "{$year}-{$month}-10" );
+		if ( ! $datetime ) {
+			return '';
+		}
+		return self::get_badge_id_from_date( $datetime->modify( 'first day of next month' ) );
+	}
+
+	/**
+	 * Get the next badge that has an excess of points, going forward up to 2 months.
+	 *
+	 * @return int
+	 */
+	public function get_next_badges_excess_points() {
+		$excess_points       = 0;
+		$next_1_badge_points = 0;
+		$next_2_badge_points = 0;
+		// Get the next badge object.
+		$next_1_badge = self::get_instance_from_id( $this->get_next_badge_id() );
+		if ( $next_1_badge ) {
+			$next_1_badge_points = $next_1_badge->progress_callback( [ 'no_next_badge_points' => true ] )['points'];
+
+			$next_2_badge = self::get_instance_from_id( $next_1_badge->get_next_badge_id() );
+			if ( $next_2_badge ) {
+				$next_2_badge_points = $next_2_badge->progress_callback( [ 'no_next_badge_points' => true ] )['points'];
+			}
+		}
+
+		$excess_points  = max( 0, $next_1_badge_points - self::TARGET_POINTS );
+		$excess_points += max( 0, $next_2_badge_points - 2 * self::TARGET_POINTS );
+
+		return (int) $excess_points;
 	}
 }