+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 = 1.6.0 =
 
+Enhancements:
+
+* Allow users to collect extra points for previous months' badges.
+
 Under the hood:
 
 * Ravi's Recommendations are now a custom post type.